Renamed 'index' routes module to 'incentivesTable'
This has no other impact just yet but it makes it clear this is the
routing related to the incentive information table.
This is currently the home page but navigation is subject to change.
