+++ server/routes/incentivesTable.ts
